Local SeriesStorage is pretty fast on 1 Billion records
Count: 1000000000
Writes, Mops: 1.83288488749753
Reads, Mops: 5.31940358846966
Count in SQLite: 1000000000
Memory size: 16000000000; SQLite net blob size: 3671028038; comp ratio: 4.36

+++ src/Spreads.Extensions/Storage/Tables.cs
@@ -10,12 +10,13 @@ namespace Spreads.Storage
     }
 
 
-
     public class SeriesChunk {
         public long Id { get; set; }
         public long ChunkKey { get; set; }
-        public int Count { get; set; }
-        public int Version { get; set; }
+        public long Count { get; set; }
+        public long Version { get; set; }
         public byte[] ChunkValue { get; set; }
     }
+
+
 }
